feat: refactor Lambda handler and update imports to use blocks
